requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-WTF==1.2.1
Flask-Migrate==4.0.7
Flask-SocketIO==5.3.6
Werkzeug==3.0.1
SQLAlchemy==2.0.28
python-dotenv==1.0.1
stripe==8.4.0
bcrypt==4.1.2
email-validator==2.1.1
python-dateutil==2.9.0
mysqlclient==2.2.4
gunicorn==21.2.0
Pillow==10.2.0
pandas==2.2.1
numpy==1.26.4
matplotlib==3.8.3
scikit-learn==1.4.0
joblib==1.3.2
xgboost==2.0.3
redis==5.0.1
flask-caching==2.1.0
aiohttp==3.9.1
asyncio==3.4.3
eventlet==0.33.3
schedule==1.2.0 
